include http tests

+++ FastEndpointTemplate.Domain/Entities/WeatherForecast.cs
@@ -4,6 +4,6 @@ public class WeatherForecast
 {
     public Guid Id { get; set; }
     public DateTime Date { get; set; }
-    public double TemperatureCelsius { get; set; }
+    public decimal TemperatureCelsius { get; set; }
     public string? Summary { get; set; }
 }
