+++ FastEndpointTemplate.Shared/Extensions/TemperatureExtension.cs
@@ -2,13 +2,12 @@ namespace FastEndpointTemplate.Shared.Extensions;
 
 public static class TemperatureExtension
 {
-    public static double ToCelsius(this double fahrenheit)
+    public static decimal ToCelsius(this decimal fahrenheit)
     {
-        return (fahrenheit - 32) * 5 / 9;
+        return (fahrenheit - 32m) * (5m / 9m);
     }
 
-
-    public static double ToFahrenheit(this double celsius)
+    public static decimal ToFahrenheit(this decimal celsius)
     {
         return celsius * 9 / 5 + 32;
     }
